chatBackend/requirements.txt
transformers==4.30.2
retry==0.9.2
flask
python-dotenv
llama-index-llms-ollama
outlines
